code/04_wake_words/requirements_new.txt
audioread==3.0.1
certifi==2020.6.20
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.4
comtypes==1.1.7
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
enum34==1.1.10
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
loguru==0.5.3
MarkupSafe==2.1.5
matplotlib==3.7.5
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.0
numba==0.58.1
numpy==1.24.4
packaging==24.1
pillow==10.2.0
platformdirs==4.3.6
pooch==1.8.2
pvporcupine==1.9.5
PyAudio @ file:///C:/Users/taffe/Documents/GitHub/AISpeechAssistant/code/04_wake_words/PyAudio-0.2.11-cp38-cp38-win_amd64.whl#sha256=d9a83a432efbe8a49f561e5ddf5bfc40e9f475386eed395e22ab3d7ff980a3f6
pycparser==2.22
pyparsing==3.1.4
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.71
pywin32==228
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.5.0
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
typeguard==4.3.0
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
webrtcvad==2.0.10
win32-setctime==1.0.3
wincertstore==0.2
zipp==3.20.2
